+++ Ansys/AnsysImport.cpp
@@ -133,8 +133,8 @@ bool AnsysImport::BuildMesh(FEModel &fem)
 	int i;
 
 	// calculate the mesh size
-	int nodes = m_Node.size();
-	int elems = m_Elem.size();
+	int nodes = (int)m_Node.size();
+	int elems = (int)m_Elem.size();
 
 	if (nodes == 0) return errf("FATAL ERROR: No nodal data defined in file.");
 	if (elems == 0) return errf("FATAL ERROR: No element data defined in file.");
